+++ src/main/java/com/pnu/momeet/domain/auth/controller/AuthController.java
@@ -1,10 +1,12 @@
 package com.pnu.momeet.domain.auth.controller;
 
 import com.pnu.momeet.common.exception.UnMatchedPasswordException;
+import com.pnu.momeet.common.security.details.CustomUserDetails;
 import com.pnu.momeet.domain.auth.dto.request.LoginRequest;
 import com.pnu.momeet.domain.auth.dto.request.RefreshRequest;
 import com.pnu.momeet.domain.auth.dto.request.SignupRequest;
 import com.pnu.momeet.domain.auth.dto.response.TokenResponse;
+import com.pnu.momeet.domain.auth.service.DefaultAuthService;
 import com.pnu.momeet.domain.auth.service.EmailAuthService;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
@@ -23,7 +25,8 @@ import java.util.UUID;
 @RequestMapping("/api/auth")
 @RequiredArgsConstructor
 public class AuthController {
-    private final EmailAuthService authService;
+    private final EmailAuthService emailAuthService;
+    private final DefaultAuthService authService;
 
     // TODO: swagger 적용하기
 
@@ -35,7 +38,7 @@ public class AuthController {
             throw new UnMatchedPasswordException("비밀번호가 일치하지 않습니다.");
         }
         TokenResponse tokenResponse =
-                authService.signUp(request.email(), request.password1());
+                emailAuthService.signUp(request.email(), request.password1());
         return ResponseEntity.ok(tokenResponse);
     }
 
@@ -43,7 +46,7 @@ public class AuthController {
     public ResponseEntity<TokenResponse> MemberLogin(
             @Valid @RequestBody LoginRequest request
     ) {
-        TokenResponse tokenResponse = authService.login(request.email(), request.password());
+        TokenResponse tokenResponse = emailAuthService.login(request.email(), request.password());
         return ResponseEntity.ok(tokenResponse);
     }
 
@@ -57,10 +60,16 @@ public class AuthController {
 
     @PostMapping("/logout")
     public ResponseEntity<Void> MemberLogout(
-            @AuthenticationPrincipal UserDetails memberInfo,
-            HttpServletResponse response
+            @AuthenticationPrincipal UserDetails memberInfo
     ) {
         authService.logout(UUID.fromString(memberInfo.getUsername()));
         return ResponseEntity.ok().build();
+    }
+
+    @PostMapping("/ws-upgrade")
+    public ResponseEntity<String> websocketUpgrade(
+            @AuthenticationPrincipal CustomUserDetails memberInfo
+    ) {
+        return ResponseEntity.ok(authService.createWsUpgradeToken(memberInfo.getMemberId()));
     }
 }
