+++ src/main/java/com/pnu/momeet/domain/auth/service/KakaoAuthService.java
@@ -1,13 +1,10 @@
 package com.pnu.momeet.domain.auth.service;
 
 import com.pnu.momeet.common.exception.BannedAccountException;
-import com.pnu.momeet.common.security.util.JwtTokenProvider;
 import com.pnu.momeet.domain.auth.client.KakaoApiClient;
 import com.pnu.momeet.domain.auth.dto.KakaoUserInfo;
 import com.pnu.momeet.domain.auth.dto.response.KakaoTokenResponse;
 import com.pnu.momeet.domain.auth.dto.response.TokenResponse;
-import com.pnu.momeet.domain.auth.entity.RefreshToken;
-import com.pnu.momeet.domain.auth.repository.RefreshTokenRepository;
 import com.pnu.momeet.domain.member.entity.Member;
 import com.pnu.momeet.domain.member.enums.Provider;
 import com.pnu.momeet.domain.member.enums.Role;
@@ -18,9 +15,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.UUID;
 
@@ -28,12 +23,8 @@ import java.util.UUID;
 @Service
 @RequiredArgsConstructor
 public class KakaoAuthService {
-
-    private static final long IAT_BUFFER_SECONDS = 10;
-
+    private final DefaultAuthService defaultAuthService;
     private final MemberEntityService memberEntityService;
-    private final JwtTokenProvider jwtTokenProvider;
-    private final RefreshTokenRepository refreshTokenRepository;
     private final KakaoApiClient kakaoApiClient;
 
     public String getKakaoAuthUrl() {
@@ -49,16 +40,13 @@ public class KakaoAuthService {
 
     public TokenResponse kakaoLogin(String code) {
         KakaoUserInfo kakaoUserInfo = getKakaoUserInfo(code);
-
         UUID memberId = findOrCreateKakaoMember(kakaoUserInfo);
-
-        TokenResponse tokenResponse = generateKakaoTokenPair(memberId);
+        TokenResponse tokenResponse = defaultAuthService.generateTokenPair(memberId);
 
         log.info("카카오 로그인 성공: {}", kakaoUserInfo.email());
         return tokenResponse;
     }
 
-    @Transactional
     private UUID findOrCreateKakaoMember(KakaoUserInfo kakaoUserInfo) {
         try {
             Member existingMember = memberEntityService.getByEmail(kakaoUserInfo.email());
@@ -85,22 +73,6 @@ public class KakaoAuthService {
         );
         log.info("카카오 회원가입 성공: {}", kakaoUserInfo.email());
         return newMember.getId();
-    }
-
-    private TokenResponse generateKakaoTokenPair(UUID memberId) {
-        Member member = memberEntityService.getById(memberId);
-
-        memberEntityService.updateMember(member, m -> {
-            m.setTokenIssuedAt(LocalDateTime.now().minusSeconds(IAT_BUFFER_SECONDS));
-            m.setEnabled(true);
-        });
-
-        String accessToken = jwtTokenProvider.generateAccessToken(memberId);
-        String refreshToken = jwtTokenProvider.generateRefreshToken(memberId);
-
-        refreshTokenRepository.save(new RefreshToken(memberId, refreshToken));
-
-        return new TokenResponse(accessToken, refreshToken);
     }
 
     @Transactional
